1. Changed the URL to vipe.t.a.i/~shlomif/freecell-solver instead of my Com-Net workstation.
2. I do not download zip.exe and unzip.exe over and over again now.


git-svn-id: file:///home/shlomif/Backup/svn-dumps/google-code/svnsync-repos/fc-solve/trunk@299 e7e8a897-7ba4-4ee7-b36f-f4c66519b19a

+++ fc-solve/site/wml/src/win32_build/bootstrap/build.bat
@@ -1,6 +1,4 @@
 REM         Let's clean the files of the previous run
-del /q zip.exe
-del /q unzip.exe
 del /q fcs.zip
 del /q freecell-solver-*.zip
 rmdir /s /q src
@@ -8,7 +6,7 @@ rmdir /s /q freecell-solver-*
 
 
 REM Assign the base URL of the site
-SET SERVER_URL=http://132.68.52.118/fc-solve/
+SET SERVER_URL=http://vipe.technion.ac.il/~shlomif/freecell-solver/
 
 REM Assign the relative place where all the files required for
 REM Building the Win32 package can be found
@@ -23,9 +21,10 @@ REM Retrieve the Version Number from the Web
 REM curl -o ver.txt %SERVER_W32_BUILD_URL%dynamic/ver.txt
 
 
-REM Fetch zip.exe and unzip.exe
-curl -o zip.exe %SERVER_W32_BUILD_URL%static/zip.exe
-curl -o unzip.exe %SERVER_W32_BUILD_URL%static/unzip.exe
+REM           Fetch zip.exe and unzip.exe 
+REM           If they were not already fetched.
+IF NOT EXIST zip.exe curl -o zip.exe %SERVER_W32_BUILD_URL%static/zip.exe
+IF NOT EXIST unzip.exe curl -o unzip.exe %SERVER_W32_BUILD_URL%static/unzip.exe
 
 REM Retrieve the archive
 curl -o fcs.zip %SERVER_W32_BUILD_URL%dynamic/fcs.zip
@@ -50,16 +49,10 @@ cd ..\..\
 REM Put the version number inside the variable VERSION
 FOR /F " tokens=1 delims=, " %%i in (.\src\ver.txt) do set VERSION=%%i
 
-REM           Prepare a directory to contain the distribution files
-REM           and put them there.
-
 SET ARC_DIR=freecell-solver-%VERSION%-bin
 mkdir %ARC_DIR%
 FOR %%i in (ChangeLog fc-solve.exe freecell-solver-range-parallel-solve.exe README README.win32.txt TODO USAGE) do copy src\%%i %ARC_DIR%\
 mkdir %ARC_DIR%\board_gen
 FOR %%i in (COPYING make_pysol_freecell_board.py make-aisleriot-freecell-board.exe make-gnome-freecell-board.exe make-microsoft-freecell-board.exe README) do copy src\board_gen\%%i %ARC_DIR%\board_gen
-
-
-REM           Prepare a zip file
 
 zip -r freecell-solver-%VERSION%.win32.bin.zip %ARC_DIR%
